# docs/openapi.front.yaml
openapi: 3.1.0
info:
  title: Lianchuan Storefront API
  version: 0.1.0
  summary: 前台电商网站接口契约
  description: |
    本文档定义前台商城所需的 REST API 契约，覆盖首页、分类、搜索、商品详情、购物车、结账、地址、订单、认证、会员资料、询价、收藏等能力。
    该文档是前台与服务端联调的唯一接口真源。
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://example.com
    description: Production placeholder
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
  parameters:
...
    OrderNumberParam:
      name: orderNumber
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type:
            - array
            - 'null'
          items:
            type: object
            additionalProperties: true
      required: [code, message]
      additionalProperties: false
...
    Money:
      type: object
      properties:
        currency:
          type: string
          example: USD
        amount:
          type: number
        formatted:
          type: string
      required: [currency, amount, formatted]
      additionalProperties: false
    ImageAsset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      required: [id, url, alt]
      additionalProperties: false
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logo:
          $ref: '#/components/schemas/ImageAsset'
      required: [id, name, slug]
      additionalProperties: false
...
    ProductCard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        shortDescription:
          type: string
        coverImage:
          $ref: '#/components/schemas/ImageAsset'
        price:
          $ref: '#/components/schemas/Money'
        compareAtPrice:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        purchaseMode:
          type: string
          enum: [buy, inquiry]
        inStock:
          type: boolean
        brand:
          oneOf:
            - $ref: '#/components/schemas/Brand'
            - type: 'null'
      required: [id, name, slug, sku, purchaseMode, inStock]
      additionalProperties: false
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        product:
          $ref: '#/components/schemas/ProductCard'
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          $ref: '#/components/schemas/Money'
        subtotal:
          $ref: '#/components/schemas/Money'
        variantLabel:
          type: string
      required: [id, productId, product, quantity, unitPrice, subtotal]
      additionalProperties: false
    Cart:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        itemCount:
          type: integer
        couponCode:
          type:
            - string
            - 'null'
        coupon:
          type:
            - object
            - 'null'
          properties:
            code:
              type: string
            description:
              type: string
            isApplied:
              type: boolean
            message:
              type:
                - string
                - 'null'
          required: [code, description, isApplied, message]
        subtotal:
          $ref: '#/components/schemas/Money'
        discount:
          $ref: '#/components/schemas/Money'
        shipping:
          $ref: '#/components/schemas/Money'
        tax:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        freeShippingThreshold:
          $ref: '#/components/schemas/Money'
        remainingForFreeShipping:
          $ref: '#/components/schemas/Money'
      required: [id, items, itemCount, subtotal, discount, shipping, tax, total, freeShippingThreshold, remainingForFreeShipping]
      additionalProperties: false
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phone:
          type: string
        countryCode:
          type: string
        state:
          type: string
        city:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
      required: [id, firstName, lastName, countryCode, city, addressLine1, postalCode, isDefault]
      additionalProperties: false
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        unitPrice:
          $ref: '#/components/schemas/Money'
        subtotal:
          $ref: '#/components/schemas/Money'
      required: [id, productName, sku, quantity, unitPrice, subtotal]
      additionalProperties: false
    OrderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        status:
          type: string
        total:
          $ref: '#/components/schemas/Money'
        createdAt:
          type: string
          format: date-time
      required: [id, orderNumber, status, total, createdAt]
      additionalProperties: false
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            shippingAddress:
              $ref: '#/components/schemas/Address'
            billingAddress:
              $ref: '#/components/schemas/Address'
          required: [items, shippingAddress, billingAddress]
          additionalProperties: false
    Inquiry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        phone:
          type: string
        company:
          type: string
        country:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, contacted, quoted, closed]
        createdAt:
          type: string
          format: date-time
      required: [id, productId, productName, email, fullName, message, status, createdAt]
      additionalProperties: false
...
    UpdateCartCouponRequest:
      type: object
      properties:
        couponCode:
          type:
            - string
            - 'null'
      additionalProperties: false
    CreateOrderRequest:
      type: object
      properties:
        shippingAddressId:
          type: string
          format: uuid
        billingAddressId:
          type: string
          format: uuid
        shippingMethod:
          type: string
        paymentMethod:
          type: string
        customerNote:
          type: string
        purchaseOrderNumber:
          type: string
        taxId:
          type: string
        requestedShipDate:
          type: string
        tradeTerm:
          type: string
      required: [shippingAddressId, billingAddressId, shippingMethod, paymentMethod]
      additionalProperties: false
    CreateInquiryRequest:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        country:
          type: string
        message:
          type: string
      required: [productId, fullName, email, message]
      additionalProperties: false
paths:
...
  /api/front/checkout/orders:
    post:
      tags: [Checkout]
      summary: 从当前购物车创建订单
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/front/orders:
    get:
      tags: [Orders]
      summary: 获取当前用户订单列表
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 订单列表
          content:
            application/json:
    patch:
      tags: [Cart]
      summary: 应用或移除购物车优惠码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartCouponRequest'
      responses:
        '200':
          description: 更新后的购物车
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: 请求非法或优惠码不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                type: array
                items:
                  $ref: '#/components/schemas/OrderSummary'
  /api/front/orders/{orderNumber}:
    get:
      tags: [Orders]
      summary: 获取订单详情
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderNumberParam'
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          description: 订单不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/front/inquiries:
    post:
      tags: [Inquiries]
      summary: 提交询价
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInquiryRequest'
      responses:
        '201':
          description: 询价提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inquiry'
    get:
      tags: [Inquiries]
      summary: 获取当前用户询价记录
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 询价列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Inquiry'
...