# src/app/api/front/checkout/orders/route.ts
import { cookies } from 'next/headers';
import { NextRequest, NextResponse } from 'next/server';
import { z } from 'zod';

import { getCurrentUserId } from '@/server/auth/session';
import { createOrderFromCart, getOrCreateCart } from '@/server/storefront/cart';

const addressSnapshotSchema = z.object({
  firstName: z.string().trim().min(1).max(80),
  lastName: z.string().trim().min(1).max(80),
  company: z.string().trim().max(120).nullable().optional(),
  phone: z.string().trim().max(60).nullable().optional(),
  countryCode: z.string().trim().min(2).max(2),
  state: z.string().trim().max(80).nullable().optional(),
  city: z.string().trim().min(1).max(120),
  addressLine1: z.string().trim().min(1).max(160),
  addressLine2: z.string().trim().max(160).nullable().optional(),
  postalCode: z.string().trim().min(1).max(32),
});

const orderSchema = z.object({
  shippingAddressId: z.string().uuid().optional(),
  billingAddressId: z.string().uuid().optional(),
  shippingAddress: addressSnapshotSchema.optional(),
  billingAddress: addressSnapshotSchema.optional(),
  shippingMethod: z.string().min(1),
  paymentMethod: z.string().min(1),
  customerNote: z.string().optional(),
  purchaseOrderNumber: z.string().max(80).optional(),
  taxId: z.string().max(80).optional(),
  requestedShipDate: z.string().max(40).optional(),
  tradeTerm: z.string().max(40).optional(),
  contactEmail: z.string().trim().email().max(160).optional(),
  subscribeToUpdates: z.boolean().optional(),
  exportComplianceConfirmed: z.boolean().optional(),
});

export async function POST(request: NextRequest) {
  const userId = await getCurrentUserId();
  const body = await request.json();
  const parsed = orderSchema.safeParse(body);
  if (!parsed.success) {
    return NextResponse.json({ code: 'VALIDATION_ERROR', message: 'Invalid checkout payload', details: parsed.error.flatten() }, { status: 400 });
  }

  const cookieStore = await cookies();
  const cart = await getOrCreateCart({ userId, anonymousToken: cookieStore.get('cart_token')?.value ?? null });
  if (!cart) {
    return NextResponse.json({ code: 'CART_UNAVAILABLE', message: 'Cart not found' }, { status: 400 });
  }

  if (!parsed.data.exportComplianceConfirmed) {
    return NextResponse.json({ code: 'EXPORT_COMPLIANCE_REQUIRED', message: 'Confirm restricted end-use compliance before placing the order.' }, { status: 400 });
  }

  if (!userId && !parsed.data.contactEmail?.trim()) {
    return NextResponse.json({ code: 'CONTACT_EMAIL_REQUIRED', message: 'Guest checkout requires a contact email.' }, { status: 400 });
  }

  const customerNote = [
    parsed.data.contactEmail ? `Contact Email: ${parsed.data.contactEmail}` : null,
    parsed.data.purchaseOrderNumber ? `PO Number: ${parsed.data.purchaseOrderNumber}` : null,
    parsed.data.taxId ? `Tax ID / VAT: ${parsed.data.taxId}` : null,
    parsed.data.requestedShipDate ? `Requested Ship Date: ${parsed.data.requestedShipDate}` : null,
    parsed.data.tradeTerm ? `Trade Term: ${parsed.data.tradeTerm}` : null,
    parsed.data.subscribeToUpdates ? 'Engineering Updates: Yes' : null,
    parsed.data.exportComplianceConfirmed ? 'Restricted End Use Confirmed: Yes' : null,
    parsed.data.customerNote?.trim() || null,
  ]
    .filter(Boolean)
    .join('\n');

  const order = userId
    ? await createOrderFromCart({
        userId,
        cartId: cart.id,
        shippingAddressId: parsed.data.shippingAddressId,
        billingAddressId: parsed.data.billingAddressId,
        shippingMethod: parsed.data.shippingMethod,
        paymentMethod: parsed.data.paymentMethod,
        customerNote,
      })
    : await createOrderFromCart({
        userId: null,
        cartId: cart.id,
        shippingAddress: parsed.data.shippingAddress,
        billingAddress: parsed.data.billingAddress,
        shippingMethod: parsed.data.shippingMethod,
        paymentMethod: parsed.data.paymentMethod,
        customerNote,
      });

  if (!order) {
    return NextResponse.json({ code: 'ORDER_CREATE_FAILED', message: 'Unable to create order' }, { status: 400 });
  }

  const redirectPath = userId ? `/account/orders/${order.orderNumber}` : `/checkout/confirmation/${order.orderNumber}`;
  const response = NextResponse.json({ orderNumber: order.orderNumber, redirectPath }, { status: 201 });

  if (!userId && 'guestAccessToken' in order && order.guestAccessToken) {
    response.cookies.set('guest_order_access_token', order.guestAccessToken, { httpOnly: true, sameSite: 'lax', path: '/' });
  }

  return response;
}